requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
setuptools==75.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
wheel==0.44.0
beautifulsoup4==4.12.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
setuptools==75.1.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
wheel==0.44.0annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.12.14
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.4.4
pydantic==2.10.4
pydantic_core==2.27.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
ollama==0.4.4
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.4
